Move constants into their own module.
